+++ src/lib/i18n/locales/en/nexus.ts
@@ -7,4 +7,7 @@ export const Strings = {
   'my-services': 'My Services',
   'back-to-services': 'Back To Nexus Services',
   'create-service': 'Create Service',
+  'use-service': 'Use This Service',
+  'import-service': 'Import Service',
+  'service-name': 'Service Name',
 } as const;
