+++ src/lib/types/nexus.ts
@@ -1,0 +1,6 @@
+export type NexusService = {
+  id: string;
+  name: string;
+  description: string;
+  tags?: string[];
+};
