+++ src/routes/(app)/nexus/[id]/import/+page.svelte
@@ -3,6 +3,8 @@
 
   import PageTitle from '$lib/components/page-title.svelte';
   import Link from '$lib/holocene/link.svelte';
+  import FilterSelect from '$lib/holocene/select/filter-select.svelte';
+  import Option from '$lib/holocene/select/simple-option.svelte';
   import { translate } from '$lib/i18n/translate';
   import { routeForNexusService } from '$lib/utilities/route-for';
 
@@ -17,15 +19,25 @@
 <div class="flex flex-col gap-8">
   <div class="relative flex flex-col gap-4 text-sm">
     <Link href={routeForNexusService(service.id)} icon="chevron-left">
-      {translate('common.view')}
-      {service.name}
+      {translate('namespaces.back-to-namespaces')}
     </Link>
   </div>
   <div class="flex flex-col gap-1">
     <div class="flex items-center justify-between">
       <h1 data-testid="namespace-selector-title" class="text-2xl">
-        {service.name}
+        {translate('nexus.import-service')}
       </h1>
     </div>
+  </div>
+  <div class="flex max-w-fit flex-col gap-4">
+    <FilterSelect
+      label={translate('nexus.service-name')}
+      parameter="name"
+      value="24 hours"
+    >
+      {#each services as service}
+        <Option value={service.name}>{service.name}</Option>
+      {/each}
+    </FilterSelect>
   </div>
 </div>
