--- src/routes/(app)/nexus/services.ts
@@ -1,79 +1,0 @@
-type Service = {
-  id: string;
-  name: string;
-  description: string;
-  tags?: string[];
-};
-
-export const services: Service[] = [
-  {
-    id: '1',
-    name: 'Activity Retry',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '2',
-    name: 'Money Transfer',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '3',
-    name: 'Saga',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '4',
-    name: 'Sleep Await',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '5',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '6',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '7',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '8',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '9',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-  {
-    id: '10',
-    name: '<Service>',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae consequat felis, vel ultrices metus. Aenean pharetra varius elit, eu elementum mauris viverra a. Vivamus non nulla at felis hendrerit fermentum ac in nisi. Nullam ut tortor eu erat scelerisque tristique. Sed egestas, magna sed ultrices commodo, sapien massa eleifend eros, vel rutrum arcu turpis eu mi.',
-    tags: ['sample', 'go', '<tag>'],
-  },
-];
